Add initial edit command
The edit command will marshal the desired resource into YAML, save it to
a file, open $EDITOR, then read in the and unmarshal the result back
into a new struct. The current implementation will do a delete then save
to update, however it would be better to add an update command.

Also, edits are not validated and user input will be directly entered into
the database.