Add support for script variables
This includes a temporary workaround to the e.GetVariables() function.
At this point, that function call will always go to the store and
populate new Variable structs. Thus, operating on the variables before
calling e.ReplaceVariables will have no effect.

The workaround is to save (update) the modified Variable to the store.
Caching should solve this issue as well and would be a better solution
so we aren't making so many calls to the store.