Add support for request variables
In order to support request variables, I also had to change the Run and
RunEnv functions to return a *http.Response which resulted in removing
the flags argument.

Also, an infinite loop formed where GenerateVariable called Run which
called GenerateVarible (and so on). This commit includes a temporary fix
via a duplicated internal run function that will not generate
RequestType variables. This will also fail a request containing a
variable that depends on a request that depends on another request
variable (i.e. run req1 ---> gen var1 ---> run req2 -X->  gen var2).